+++ bindings/wasm/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Not respecting specificity in case of inlining overlapping rules like `padding` and `padding-top`. [#142](https://github.com/Stranger6667/css-inline/issues/142)
+- `!important` rules not overriding inlined styles. [#152](https://github.com/Stranger6667/css-inline/issues/152)
 
 ## [0.8.0] - 2022-01-10